# gateway (copy)/docs/swagger.yaml
basePath: /
definitions:
  genproto.AccounList:
    properties:
      Accounts:
        items:
          $ref: '#/definitions/genproto.AccountGet'
        type: array
      Limit:
        type: integer
      Offset:
        type: integer
      TotalCount:
        type: integer
    type: object
  genproto.AccountCreate:
    properties:
      Currency:
        type: string
      Name:
        type: string
      Type:
        type: string
      UserId:
        type: string
    type: object
  genproto.AccountGet:
    properties:
      CreatedAt:
        type: string
      Id:
        type: string
      UpdatedAt:
        type: string
      UserId:
        type: string
      balance:
        type: number
      currency:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  genproto.AccountUpt:
    properties:
      balance:
        type: number
      currency:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
  /account/{id}:
    get:
      consumes:
      - application/json
      description: Get an account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account data
          schema:
            $ref: '#/definitions/genproto.AccountGet'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Account not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Account
      tags:
      - Account
  /account/create:
    post:
      consumes:
      - application/json
      description: Create a new account
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/genproto.AccountCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Account created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Account
      tags:
      - Account
  /account/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Account
      tags:
      - Account
  /account/list:
    get:
      consumes:
      - application/json
      description: List accounts with filters
      parameters:
      - description: Account Name
        in: query
        name: name
        type: string
      - description: Account Type
        in: query
        name: type
        type: string
      - description: Currency
        in: query
        name: currency
        type: string
      - description: Minimum Balance
        in: query
        name: balanceMin
        type: integer
      - description: Maximum Balance
        in: query
        name: balanceMax
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: User ID
        in: query
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of accounts
          schema:
            $ref: '#/definitions/genproto.AccounList'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Accounts
      tags:
      - Account
  /account/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/genproto.AccountUpt'
      produces:
      - application/json
      responses:
        "200":
          description: Account updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Account
      tags:
      - Account
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"